+++ src/main/java/com/redhat/lightblue/camel/response/LightblueResponseTransformer.java
@@ -1,4 +1,4 @@
-package com.redhat.lightblue.camel.utils;
+package com.redhat.lightblue.camel.response;
 
 import org.apache.camel.Handler;
